+++ stream-chat-android-client/src/main/java/io/getstream/chat/android/client/notifications/handler/NotificationConfig.kt
@@ -85,4 +85,9 @@ public data class NotificationConfig(
      * @see [PushDeviceGenerator]
      */
     val pushDeviceGenerators: List<PushDeviceGenerator> = listOf(),
+
+    /**
+     * Allows enabling/disabling showing notification after receiving a push message.
+     */
+    val shouldShowNotificationOnPush: () -> Boolean = { true },
 )